requirements.txt
python-dotenv==1.0.0
retell-sdk==1.0.0
pydantic>=1.10,<2.0
openai>=1.0.0
elevenlabs==0.2.24
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0
python-multipart==0.0.6
PyYAML==6.0.1
sounddevice==0.4.6
numpy==1.26.2
requests==2.31.0
aiohttp==3.9.1
asyncio==3.4.3
loguru==0.7.2
redis==5.0.1
aioredis==2.0.1
prometheus-client==0.19.0
gunicorn==21.2.0
deepgram-sdk==2.11.0

# Testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
